china mirror notification updated to a popover
